requirements.txt
requests==2.31.0
python-dotenv==1.0.0
slack-sdk==3.26.0
discord.py==2.3.2
schedule==1.2.0
pyyaml==6.0.1


